Bluetooth: Controller: Rename hci_le_rand to bt_rand
Rename hci_le_rand to bt_rand and make its parameter types compatible.
This also includes updating rand_get() to use size_t instead of
uint32_t & uint8_t.


Former-commit-id: 815e15e3b2f7c839cbeca02e0e468e5cf85b6d77

+++ subsys/bluetooth/controller/hal/nrf5/rand.c
@@ -3,7 +3,7 @@ DECL|count|member|uint8_t count;
 DECL|first|member|uint8_t first;
 DECL|isr_rand|function|void isr_rand(void *param)
 DECL|last|member|uint8_t last;
-DECL|rand_get|function|uint32_t rand_get(uint8_t octets, uint8_t *rand)
+DECL|rand_get|function|size_t rand_get(size_t octets, uint8_t *rand)
 DECL|rand_init|function|void rand_init(uint8_t *context, uint8_t context_len)
 DECL|rand|member|uint8_t rand[1];
 DECL|rand|struct|struct rand {
